requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-filter
django-extensions
django-cors-headers
gunicorn
whitenoise
python-dotenv
pytz
requests
redis
yt-dlp
yt-dlp-ejs
ffmpeg-python
django-rq